requirements.txt
discord.py
matplotlib
seaborn
circlify
colorthief
pillow
pandas
numpy